requirement.txt
Django
djangorestframework
sklearn
pandas 
numpy
matplotlib
seaborn
gunicorn